AST-based ns form refactoring (#2)
* Begin work on parsing NS forms via AST

* Fix newline handling + support vectors

* Track AST indent level + infer indent amount for Append

* Add test to verify indent behavior with vectors on their own line

* Add parent links to AST nodes

* Support inserting literals into vectors

* Support inserting keypairs to :require vectors

* Do sorted insert into forms as well, not just vectors

Fixes adding new :require entries. May need to a special case
in s:compare to support appending :import *after* :require...

* Improve form's sorted insert algorithm

* Support inserting into nested forms

All old tests now pass with the new AST-based code!
Still TODO: infer use of [] vs () for :refer
Eventually: support :import

* Infer :refer style (brackets vs parens)

* Remove unused .parent property + add context to Throw()

* Match indentation on prototypes

* Refactor Vector.SortedAddKeyPair to use newer list-extend style

* Simplify AST interface by removing the Sorted descriptor on methods

We don't have any non-sorted operations, since everything is intended to
be non-obtrusive and part of a natural edit, so everything should be
assumed to be sorted. Including that information in the method name is
redundant and annoying to read.

* Handle namespaces with commented bits

* Share insert index computation

* Refactor for simplicity, organization, uniformity